Initial refactoring for compiler API
