+++ src/compiler/utils/recordFieldReferenceTable.ts
