+++ src/compiler/utils/syntaxNodeMap.ts
