+++ src/providers/codeAction/extractFunctionCodeAction.ts
@@ -2,7 +2,7 @@ import { CodeActionKind, Position, TextEdit } from "vscode-languageserver";
 import { SyntaxNode } from "web-tree-sitter";
 import { RefactorEditUtils } from "../../util/refactorEditUtils";
 import { TreeUtils } from "../../util/treeUtils";
-import { TFunction } from "../../util/types/typeInference";
+import { TFunction } from "../../compiler/typeInference";
 import {
   CodeActionProvider,
   IRefactorCodeAction,
