+++ src/providers/codeAction/installPackageCodeAction.ts
@@ -3,12 +3,12 @@ import { container } from "tsyringe";
 import { CodeAction, Connection } from "vscode-languageserver";
 import { URI } from "vscode-uri";
 import { CommandManager } from "../../commandManager";
-import { ElmPackageCache } from "../../elmPackageCache";
-import { execCmdSync } from "../../util/elmUtils";
+import { ElmPackageCache } from "../../compiler/elmPackageCache";
+import { execCmdSync } from "../../compiler/utils/elmUtils";
 import { ElmWorkspaceMatcher } from "../../util/elmWorkspaceMatcher";
 import { Settings } from "../../util/settings";
 import { TreeUtils } from "../../util/treeUtils";
-import { Diagnostics } from "../../util/types/diagnostics";
+import { Diagnostics } from "../../compiler/diagnostics";
 import { CodeActionProvider } from "../codeActionProvider";
 import { ICodeActionParams } from "../paramsExtensions";
 import { comparePackageRanking } from "../ranking";
