+++ src/providers/documentFormatingProvider.ts
@@ -7,7 +7,7 @@ import {
 import { URI } from "vscode-uri";
 import { DiagnosticsProvider } from ".";
 import * as Diff from "../util/diff";
-import { execCmdSync } from "../util/elmUtils";
+import { execCmdSync } from "../compiler/utils/elmUtils";
 import { ElmWorkspaceMatcher } from "../util/elmWorkspaceMatcher";
 import { Settings } from "../util/settings";
 import { TextDocumentEvents } from "../util/textDocumentEvents";
