+++ src/util/renameUtils.ts
@@ -1,7 +1,7 @@
-import { IProgram } from "../program";
+import { IProgram } from "../compiler/program";
 import { Position, ResponseError } from "vscode-languageserver";
 import { SyntaxNode } from "web-tree-sitter";
-import { References } from "./references";
+import { References } from "../compiler/references";
 import { TreeUtils } from "./treeUtils";
 
 export class RenameUtils {
