+++ test/diagnostics.ts
@@ -1,13 +1,13 @@
 import "reflect-metadata";
 import { container } from "tsyringe";
 import { URI } from "vscode-uri";
-import { Program } from "../src/program";
+import { Program } from "../src/compiler/program";
 import * as path from "path";
 import { Settings } from "../src/util/settings";
 import Parser from "web-tree-sitter";
 import { spawnSync } from "child_process";
 import { readFileSync } from "fs";
-import { Diagnostic } from "../src/util/types/diagnostics";
+import { Diagnostic } from "../src/compiler/diagnostics";
 import { performance } from "perf_hooks";
 import { createProgramHost } from "./utils/sourceTreeParser";
